% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app_server}
\alias{app_server}
\title{App Server}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}
}
\description{
Main server function handling module initialization,
performance monitoring, and user interactions.
}
