% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CACHE_MAX_AGE_SECONDS}
\alias{CACHE_MAX_AGE_SECONDS}
\title{Cache maximum age in seconds}
\format{
An object of class \code{integer} of length 1.
}
\usage{
CACHE_MAX_AGE_SECONDS
}
\description{
Cache maximum age in seconds
}
\keyword{datasets}
