% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CACHE_MAX_SIZE_MB}
\alias{CACHE_MAX_SIZE_MB}
\title{Cache maximum size in MB}
\format{
An object of class \code{integer} of length 1.
}
\usage{
CACHE_MAX_SIZE_MB
}
\description{
Cache maximum size in MB
}
\keyword{datasets}
