% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiral_math.R
\name{compute_voronoi}
\alias{compute_voronoi}
\title{Compute Voronoi Diagram}
\usage{
compute_voronoi(spiral_points)
}
\arguments{
\item{spiral_points}{Matrix with x,y coordinates}
}
\value{
List with voronoi diagram and bounded count
}
\description{
Compute Voronoi Diagram
}
