% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{create_performance_report}
\alias{create_performance_report}
\title{Create Performance Report}
\usage{
create_performance_report(timings)
}
\arguments{
\item{timings}{Numeric vector of timing values (spiral, voronoi, plot)}
}
\value{
Performance report object
}
\description{
Create Performance Report
}
