% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEBOUNCE_MS}
\alias{DEBOUNCE_MS}
\title{Debounce delay in milliseconds for slider inputs}
\format{
An object of class \code{integer} of length 1.
}
\usage{
DEBOUNCE_MS
}
\description{
Debounce delay in milliseconds for slider inputs
}
\keyword{datasets}
