% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_PALETTE}
\alias{DEFAULT_PALETTE}
\title{Default color palette}
\format{
An object of class \code{character} of length 1.
}
\usage{
DEFAULT_PALETTE
}
\description{
Default color palette
}
\keyword{datasets}
