% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_PLOT_LIMITS}
\alias{DEFAULT_PLOT_LIMITS}
\title{Default plot limits when calculation fails}
\format{
An object of class \code{integer} of length 2.
}
\usage{
DEFAULT_PLOT_LIMITS
}
\description{
Default plot limits when calculation fails
}
\keyword{datasets}
