% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{.defaults}
\alias{.defaults}
\title{Default configuration values (fallbacks when config.yml missing)}
\format{
An object of class \code{list} of length 10.
}
\usage{
.defaults
}
\description{
Default configuration values (fallbacks when config.yml missing)
}
\keyword{internal}
