% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_manager.R
\name{.get_compute_fn}
\alias{.get_compute_fn}
\title{Get or create the memoized compute function}
\usage{
.get_compute_fn()
}
\description{
Get or create the memoized compute function
}
\keyword{internal}
