% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{.load_config}
\alias{.load_config}
\title{Load configuration from config.yml}
\usage{
.load_config()
}
\value{
List with configuration values or NULL if not found
}
\description{
Load configuration from config.yml
}
\keyword{internal}
