% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiral_math.R
\name{estimate_computation_time}
\alias{estimate_computation_time}
\title{Estimate Computation Time}
\usage{
estimate_computation_time(num_points)
}
\arguments{
\item{num_points}{Number of points}
}
\value{
Estimated time in milliseconds
}
\description{
Estimate Computation Time
}
