% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{EXPORT_PNG_RES}
\alias{EXPORT_PNG_RES}
\title{PNG export resolution (DPI)}
\format{
An object of class \code{integer} of length 1.
}
\usage{
EXPORT_PNG_RES
}
\description{
PNG export resolution (DPI)
}
\keyword{datasets}
