% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{EXPORT_PNG_SIZE}
\alias{EXPORT_PNG_SIZE}
\title{PNG export dimensions (pixels)}
\format{
An object of class \code{integer} of length 1.
}
\usage{
EXPORT_PNG_SIZE
}
\description{
PNG export dimensions (pixels)
}
\keyword{datasets}
