% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{EXPORT_SVG_SIZE}
\alias{EXPORT_SVG_SIZE}
\title{SVG export dimensions (inches)}
\format{
An object of class \code{integer} of length 1.
}
\usage{
EXPORT_SVG_SIZE
}
\description{
SVG export dimensions (inches)
}
\keyword{datasets}
