% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{get_app_dir}
\alias{get_app_dir}
\title{Get App Directory}
\usage{
get_app_dir()
}
\value{
Character path to app directory
}
\description{
Returns the path to the installed Shiny app directory.
}
\keyword{internal}
