% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{get_spiralizer_theme}
\alias{get_spiralizer_theme}
\title{Get the Spiralizer theme}
\usage{
get_spiralizer_theme()
}
\value{
bslib theme object
}
\description{
Returns the theme object. Use this function if you need
to modify or extend the theme.
}
