% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-utils.R
\name{null-coalesce}
\alias{null-coalesce}
\alias{\%||\%}
\title{Null coalescing operator}
\usage{
lhs \%||\% rhs
}
\arguments{
\item{lhs}{Left-hand side value}

\item{rhs}{Right-hand side (default) value}
}
\value{
lhs if not NULL, otherwise rhs
}
\description{
Returns lhs if not NULL, otherwise rhs. Equivalent to rlang's \%||\%.
}
\keyword{internal}
