% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{PLOT_LIMIT_PADDING}
\alias{PLOT_LIMIT_PADDING}
\title{Plot limit padding factor (10\% padding)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
PLOT_LIMIT_PADDING
}
\description{
Plot limit padding factor (10\% padding)
}
\keyword{datasets}
