% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{preview_theme}
\alias{preview_theme}
\title{Preview the theme in an interactive viewer}
\usage{
preview_theme()
}
\description{
Opens an interactive bslib theme preview for development.
Allows real-time color adjustments.
}
