% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SLIDER_ANGLE_MAX}
\alias{SLIDER_ANGLE_MAX}
\title{Slider maximum for angles}
\format{
An object of class \code{integer} of length 1.
}
\usage{
SLIDER_ANGLE_MAX
}
\description{
Slider maximum for angles
}
\keyword{datasets}
