% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SLIDER_ANGLE_MIN}
\alias{SLIDER_ANGLE_MIN}
\title{Slider minimum for angles}
\format{
An object of class \code{integer} of length 1.
}
\usage{
SLIDER_ANGLE_MIN
}
\description{
Slider minimum for angles
}
\keyword{datasets}
