% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SLIDER_DENSITY_MAX}
\alias{SLIDER_DENSITY_MAX}
\title{Slider maximum for density}
\format{
An object of class \code{integer} of length 1.
}
\usage{
SLIDER_DENSITY_MAX
}
\description{
Slider maximum for density
}
\keyword{datasets}
