% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SLIDER_DENSITY_MIN}
\alias{SLIDER_DENSITY_MIN}
\title{Slider minimum for density}
\format{
An object of class \code{integer} of length 1.
}
\usage{
SLIDER_DENSITY_MIN
}
\description{
Slider minimum for density
}
\keyword{datasets}
