% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SPIRAL_DEFAULT_POINTS}
\alias{SPIRAL_DEFAULT_POINTS}
\title{Default number of points}
\format{
An object of class \code{integer} of length 1.
}
\usage{
SPIRAL_DEFAULT_POINTS
}
\description{
Default number of points
}
\keyword{datasets}
