% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SPIRAL_MAX_ANGLE_RANGE}
\alias{SPIRAL_MAX_ANGLE_RANGE}
\title{Maximum angle range allowed}
\format{
An object of class \code{integer} of length 1.
}
\usage{
SPIRAL_MAX_ANGLE_RANGE
}
\description{
Maximum angle range allowed
}
\keyword{datasets}
