% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SPIRAL_MAX_POINTS}
\alias{SPIRAL_MAX_POINTS}
\title{Maximum number of points allowed for spiral generation}
\format{
An object of class \code{integer} of length 1.
}
\usage{
SPIRAL_MAX_POINTS
}
\description{
Maximum number of points allowed for spiral generation
}
\keyword{datasets}
