% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SPIRAL_MIN_POINTS}
\alias{SPIRAL_MIN_POINTS}
\title{Minimum number of points required for Voronoi diagram}
\format{
An object of class \code{integer} of length 1.
}
\usage{
SPIRAL_MIN_POINTS
}
\description{
Minimum number of points required for Voronoi diagram
}
\keyword{datasets}
