% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_zen.R
\name{spiralizer_app}
\alias{spiralizer_app}
\title{Create Spiralizer App Object}
\usage{
spiralizer_app()
}
\value{
A Shiny app object
}
\description{
Returns a Shiny app object that can be run with \code{shiny::runApp()}.
}
