% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\docType{data}
\name{spiralizer_theme}
\alias{spiralizer_theme}
\title{Spiralizer Theme}
\format{
An object of class \code{bs_theme_with_preset} (inherits from \code{bs_version_5}, \code{bs_theme}, \code{sass_bundle}) of length 1.
}
\usage{
spiralizer_theme
}
\value{
A bslib theme object
}
\description{
Creates the main application theme using bslib.
Based on Bootstrap 5 with darkly preset and zen customizations.
}
\keyword{datasets}
