% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\docType{data}
\name{theme_colors}
\alias{theme_colors}
\title{Theme color palette}
\format{
An object of class \code{list} of length 8.
}
\usage{
theme_colors
}
\description{
Named list of theme colors for use in R plotting code.
}
\keyword{datasets}
