% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiral_math.R
\name{validate_spiral_params}
\alias{validate_spiral_params}
\title{Validate Spiral Parameters}
\usage{
validate_spiral_params(angle_start, angle_end, num_points)
}
\arguments{
\item{angle_start}{Starting angle}

\item{angle_end}{Ending angle}

\item{num_points}{Number of points}
}
\value{
List with valid flag and message
}
\description{
Validate Spiral Parameters
}
