% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\docType{data}
\name{zen_colors}
\alias{zen_colors}
\title{Zen color palette}
\format{
An object of class \code{list} of length 8.
}
\usage{
zen_colors
}
\description{
Named list of zen theme colors for use in R plotting code.
}
\keyword{datasets}
