% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_zen.R
\name{zen_server}
\alias{zen_server}
\title{Zen Server}
\usage{
zen_server(input, output, session)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}
}
\description{
Main server function handling module initialization,
performance monitoring, and user interactions.
}
