% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_utils.R
\name{get_color_palette}
\alias{get_color_palette}
\title{Get Color Palette}
\usage{
get_color_palette(palette_name, n_colors, invert = FALSE)
}
\arguments{
\item{palette_name}{Character name of the palette}

\item{n_colors}{Number of colors to generate}

\item{invert}{Logical, whether to reverse the palette}
}
\value{
Character vector of hex colors
}
\description{
Returns a color vector for the specified palette.
Centralizes palette logic to avoid duplication in plot rendering.
}
