% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\docType{data}
\name{palette_choices}
\alias{palette_choices}
\title{Available color palettes for spirals}
\format{
An object of class \code{list} of length 7.
}
\usage{
palette_choices
}
\description{
Named list for use in selectInput choices.
}
\keyword{datasets}
